Bullet 이 생성/재활용 될때 다른 power 를 가지고 생성되도록.

+++ projects/DragonFlight2/app/src/main/java/kr/ac/tukorea/ge/sgp02/s12345678/dragonflight/game/Enemy.java
@@ -20,7 +20,7 @@ public class Enemy extends AnimSprite implements BoxCollidable, Recyclable {
     private static final String TAG = Enemy.class.getSimpleName();
     public static float size;
     protected int level;
-    protected int life, maxLife;
+    protected float life, maxLife;
     protected Gauge gauge;
     protected float dy;
     protected RectF boundingBox = new RectF();
@@ -104,7 +104,7 @@ public class Enemy extends AnimSprite implements BoxCollidable, Recyclable {
         return level * level * 100;
     }
 
-    public boolean decreaseLife(int power) {
+    public boolean decreaseLife(float power) {
         life -= power;
         if (life <= 0) return true;
